requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
bcrypt==4.3.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.2
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
packaging==25.0
phonenumbers==9.0.14
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
uvloop==0.21.0
validate_docbr==1.11.1
watchfiles==1.1.0
websockets==15.0.1
